:- ** BRANCH: compile_fact ** -> ** FILE SIZE ERROR FIX ATTEMPT ** -> ** MIGRATED _getMarketForTicket to LIB ** 	- updated function signature and now receives & returns memory reference instead of storage 	- invoked from FACTORY in various functions 		also write back to storage when any memory references are changed 			in -> ‘setMarketInfo’, ‘castVoteForMarketTicket’, and ‘closeMarketForTicket’
:- NOTE: this was an attempt to lower the FACTORY file size from 24638 to 24576
	HOWEVER, it appears that it was raised to 27428
	
:- NOTE: current errors
Warning: Contract code size is 27428 bytes and exceeds 24576 bytes (a limit introduced in Spurious Dragon). This contract may not be deployable on Mainnet. Consider enabling the optimizer (with a low "runs" value!), turning off revert strings, or using libraries.

:- NOTE: was…
Warning: Contract code size is 24638 bytes and exceeds 24576 bytes (a limit introduced in Spurious Dragon). This contract may not be deployable on Mainnet. Consider enabling the optimizer (with a low "runs" value!), turning off revert strings, or using libraries.